# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: documents
    description: Contract workspace operations
  - name: questions
    description: Grounded document questions
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List uploaded contract workspaces
      responses:
        "200":
          description: Documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DocumentSummary"
    post:
      operationId: createDocument
      tags: [documents]
      summary: Create a contract workspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentInput"
      responses:
        "201":
          description: Created document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentSummary"
  /documents/upload:
    post:
      operationId: uploadDocument
      tags: [documents]
      summary: Upload and shield a contract document
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required: [file]
      responses:
        "201":
          description: Parsed and shielded document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentUploadResponse"
        "422":
          description: Scanned document rejected
  /documents/{id}:
    get:
      operationId: getDocument
      tags: [documents]
      summary: Get a contract workspace with its clauses
      parameters:
        - $ref: "#/components/parameters/DocumentId"
      responses:
        "200":
          description: Document details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentDetail"
        "404":
          description: Document not found
    post:
      operationId: analyzeDocument
      tags: [documents]
      summary: Start deterministic document analysis
      parameters:
        - $ref: "#/components/parameters/DocumentId"
      responses:
        "202":
          description: Analysis queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentSummary"
        "404":
          description: Document not found
  /documents/{id}/compare:
    post:
      operationId: compareDocumentVersions
      tags: [documents]
      summary: Compare two contract versions clause by clause
      description: >-
        Aligns the clauses of the document in the path (the revised version)
        against the `against` document (the base version) using heading
        similarity, clause-type match and Jaccard token similarity. Re-runs the
        deterministic risk rule pack on both versions, diffs the findings, and
        sends only the matched pairs below the similarity threshold to the
        smart model, one pair per call, cached by the pair's combined hash.
      parameters:
        - $ref: "#/components/parameters/DocumentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompareVersionsBody"
      responses:
        "200":
          description: Clause alignment, change labels and risk findings diff
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VersionComparison"
        "404":
          description: Base or revised version not found
  /documents/{id}/ask:
    post:
      operationId: askDocumentQuestion
      tags: [questions]
      summary: Ask a grounded question about a specific document
      parameters:
        - $ref: "#/components/parameters/DocumentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionInput"
      responses:
        "200":
          description: Grounded answer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnswerResponse"
            text/event-stream:
              schema:
                type: string
                description: Server-sent events stream with status, answer, and done events.
  /questions:
    post:
      operationId: askQuestion
      tags: [questions]
      summary: Ask a grounded question about a document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionInput"
      responses:
        "200":
          description: Grounded answer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnswerResponse"
components:
  parameters:
    DocumentId:
      name: id
      in: path
      required: true
      schema:
        type: string
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    DocumentInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        fileType:
          type: string
        sizeBytes:
          type: integer
          minimum: 1
      required: [name, fileType, sizeBytes]
    RiskSeverity:
      type: string
      enum: [low, medium, high]
    DocumentStatus:
      type: string
      enum: [uploaded, analyzing, ready]
    SecurityFinding:
      type: object
      properties:
        label:
          type: string
        detail:
          type: string
        page:
          type: integer
      required: [label, detail, page]
    InjectionFinding:
      type: object
      properties:
        pattern:
          type: string
        page:
          type: integer
        offset:
          type: integer
        severity:
          type: string
          enum: [info, warning]
      required: [pattern, page, offset, severity]
    DocumentSummary:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        fileType:
          type: string
        uploadedAt:
          type: string
          format: date-time
        status:
          $ref: "#/components/schemas/DocumentStatus"
        pageCount:
          type: integer
        clauseCount:
          type: integer
        riskScore:
          type: integer
        highRiskCount:
          type: integer
        scannedDetected:
          type: boolean
        securityFindings:
          type: array
          items:
            $ref: "#/components/schemas/SecurityFinding"
        injectionFindings:
          type: array
          items:
            $ref: "#/components/schemas/InjectionFinding"
      required:
        - id
        - name
        - fileType
        - uploadedAt
        - status
        - pageCount
        - clauseCount
        - riskScore
        - highRiskCount
        - scannedDetected
        - securityFindings
    Clause:
      type: object
      properties:
        id:
          type: string
        ordinal:
          type: string
        heading:
          type: string
        type:
          type: string
        severity:
          $ref: "#/components/schemas/RiskSeverity"
        summary:
          type: string
        text:
          type: string
      required: [id, ordinal, heading, type, severity, summary, text]
    DocumentDetail:
      allOf:
        - $ref: "#/components/schemas/DocumentSummary"
        - type: object
          properties:
            clauses:
              type: array
              items:
                $ref: "#/components/schemas/Clause"
          required: [clauses]
    DocumentUploadResponse:
      allOf:
        - $ref: "#/components/schemas/DocumentDetail"
        - type: object
          properties:
            expiresAt:
              type: string
              format: date-time
          required: [expiresAt]
    QuestionInput:
      type: object
      properties:
        documentId:
          type: string
        question:
          type: string
          minLength: 1
        perspective:
          type: string
          enum: [party_a, party_b]
      required: [documentId, question, perspective]
    Citation:
      type: object
      properties:
        clauseId:
          type: string
        label:
          type: string
      required: [clauseId, label]
    AnswerResponse:
      type: object
      properties:
        answer:
          type: string
        citations:
          type: array
          items:
            $ref: "#/components/schemas/Citation"
        groundingRatio:
          type: number
        adviceMode:
          type: boolean
        queued:
          type: boolean
      required: [answer, citations, groundingRatio, adviceMode, queued]
    CompareVersionsBody:
      type: object
      properties:
        against:
          type: string
          description: Id of the earlier (base) version to compare against.
        perspective:
          type: string
          enum: [both, party_a, party_b]
          description: Which side the reader is on, used to label which party a change favours.
      required: [against]
    VersionDocumentRef:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
      required: [id, name]
    ClauseRef:
      type: object
      properties:
        id:
          type: string
        ordinal:
          type: string
        heading:
          type: string
        type:
          type: string
        text:
          type: string
      required: [id, ordinal, heading, type, text]
    SimilarityBreakdown:
      type: object
      properties:
        headingSimilarity:
          type: number
        clauseTypeMatch:
          type: number
        tokenSimilarity:
          type: number
      required: [headingSimilarity, clauseTypeMatch, tokenSimilarity]
    MaterialChangeNote:
      type: object
      properties:
        available:
          type: boolean
        fromCache:
          type: boolean
        materiallyChanged:
          type: boolean
        summary:
          type: string
        favors:
          type: string
          enum: [partyA, partyB, neither]
        rationale:
          type: string
      required: [available, fromCache, materiallyChanged, summary, favors, rationale]
    FindingChange:
      type: object
      properties:
        status:
          type: string
          enum: [introduced, resolved, changed, unchanged]
        label:
          type: string
          enum: [Added, Removed, Changed]
        ruleId:
          type: string
        title:
          type: string
        severity:
          $ref: "#/components/schemas/RiskSeverity"
        concerns:
          type: string
          enum: [partyA, partyB, both]
        favorsOtherParty:
          type: boolean
        message:
          type: string
        beforeClauseId:
          type: string
        afterClauseId:
          type: string
      required: [status, ruleId, title, severity, concerns, favorsOtherParty, message]
    ComparisonRow:
      type: object
      properties:
        key:
          type: string
        kind:
          type: string
          enum: [unchanged, changed, added, removed]
        similarity:
          type: number
        labels:
          type: array
          items:
            type: string
            enum: [Added, Removed, Changed, Favors other party]
        before:
          $ref: "#/components/schemas/ClauseRef"
        after:
          $ref: "#/components/schemas/ClauseRef"
        breakdown:
          $ref: "#/components/schemas/SimilarityBreakdown"
        materialChange:
          $ref: "#/components/schemas/MaterialChangeNote"
        findingChanges:
          type: array
          items:
            $ref: "#/components/schemas/FindingChange"
      required: [key, kind, similarity, labels, findingChanges]
    VersionComparisonSummary:
      type: object
      properties:
        totalRows:
          type: integer
        changedCount:
          type: integer
        addedCount:
          type: integer
        removedCount:
          type: integer
        unchangedCount:
          type: integer
        introducedRiskCount:
          type: integer
        resolvedRiskCount:
          type: integer
        favorsOtherPartyCount:
          type: integer
      required:
        - totalRows
        - changedCount
        - addedCount
        - removedCount
        - unchangedCount
        - introducedRiskCount
        - resolvedRiskCount
        - favorsOtherPartyCount
    VersionComparison:
      type: object
      properties:
        base:
          $ref: "#/components/schemas/VersionDocumentRef"
        revised:
          $ref: "#/components/schemas/VersionDocumentRef"
        threshold:
          type: number
        perspective:
          type: string
          enum: [partyA, partyB, both]
        rows:
          type: array
          items:
            $ref: "#/components/schemas/ComparisonRow"
        summary:
          $ref: "#/components/schemas/VersionComparisonSummary"
      required: [base, revised, threshold, perspective, rows, summary]