# lib/api-zod/src/generated/types/clauseRef.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ClauseRef {
  id: string;
  ordinal: string;
  heading: string;
  type: string;
  text: string;
}