# lib/api-zod/src/generated/types/compareVersionsBody.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CompareVersionsBodyPerspective } from './compareVersionsBodyPerspective';

export interface CompareVersionsBody {
  /** Id of the earlier (base) version to compare against. */
  against: string;
  /** Which side the reader is on, used to label which party a change favours. */
  perspective?: CompareVersionsBodyPerspective;
}