# lib/api-zod/src/generated/types/comparisonRow.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ClauseRef } from './clauseRef';
import type { ComparisonRowKind } from './comparisonRowKind';
import type { ComparisonRowLabelsItem } from './comparisonRowLabelsItem';
import type { FindingChange } from './findingChange';
import type { MaterialChangeNote } from './materialChangeNote';
import type { SimilarityBreakdown } from './similarityBreakdown';

export interface ComparisonRow {
  key: string;
  kind: ComparisonRowKind;
  similarity: number;
  labels: ComparisonRowLabelsItem[];
  before?: ClauseRef;
  after?: ClauseRef;
  breakdown?: SimilarityBreakdown;
  materialChange?: MaterialChangeNote;
  findingChanges: FindingChange[];
}