# lib/api-zod/src/generated/types/findingChange.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FindingChangeConcerns } from './findingChangeConcerns';
import type { FindingChangeLabel } from './findingChangeLabel';
import type { FindingChangeStatus } from './findingChangeStatus';
import type { RiskSeverity } from './riskSeverity';

export interface FindingChange {
  status: FindingChangeStatus;
  label?: FindingChangeLabel;
  ruleId: string;
  title: string;
  severity: RiskSeverity;
  concerns: FindingChangeConcerns;
  favorsOtherParty: boolean;
  message: string;
  beforeClauseId?: string;
  afterClauseId?: string;
}