# lib/api-zod/src/generated/types/materialChangeNote.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MaterialChangeNoteFavors } from './materialChangeNoteFavors';

export interface MaterialChangeNote {
  available: boolean;
  fromCache: boolean;
  materiallyChanged: boolean;
  summary: string;
  favors: MaterialChangeNoteFavors;
  rationale: string;
}