# lib/api-zod/src/generated/types/similarityBreakdown.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SimilarityBreakdown {
  headingSimilarity: number;
  clauseTypeMatch: number;
  tokenSimilarity: number;
}