# lib/api-zod/src/generated/types/versionComparison.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ComparisonRow } from './comparisonRow';
import type { VersionComparisonPerspective } from './versionComparisonPerspective';
import type { VersionComparisonSummary } from './versionComparisonSummary';
import type { VersionDocumentRef } from './versionDocumentRef';

export interface VersionComparison {
  base: VersionDocumentRef;
  revised: VersionDocumentRef;
  threshold: number;
  perspective: VersionComparisonPerspective;
  rows: ComparisonRow[];
  summary: VersionComparisonSummary;
}