# lib/api-zod/src/generated/types/versionComparisonSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface VersionComparisonSummary {
  totalRows: number;
  changedCount: number;
  addedCount: number;
  removedCount: number;
  unchangedCount: number;
  introducedRiskCount: number;
  resolvedRiskCount: number;
  favorsOtherPartyCount: number;
}